Replace some types with interfaces to avoid relying on function.name to get the filter name
When the program is bundled with Webpack or similar tools, the function may change names or even
become anonymous, making it not possible to rely on the function.name for anything, like I did
to get the filter name when it was applied in TextTransformer.
By using an interface for Filter, instead of a type, I can define an additional property and add
that properties to the objects that implement that, or similar, interfaces (e.g. CaseConverters)